+++ blaze-client/src/main/scala/org/http4s/client/blaze/Http1Support.scala
@@ -24,34 +24,41 @@ import scalaz.concurrent.Task
 
 import scalaz.{\/, -\/, \/-}
 
+object Http1Support {
+  /** Create a new [[ConnectionBuilder]]
+   *
+   * @param bufferSize buffer size of the socket stages
+   * @param userAgent User-Agent header information
+   * @param es `ExecutorService` on which computations should be run
+   * @param osslContext Optional `SSSContext` for secure requests
+   * @param group `AsynchronousChannelGroup` used to manage socket connections
+   * @return [[ConnectionBuilder]] for creating new requests
+   */
+  def apply(bufferSize: Int,
+             userAgent: Option[`User-Agent`],
+                    es: ExecutorService,
+           osslContext: Option[SSLContext],
+                 group: Option[AsynchronousChannelGroup]): ConnectionBuilder = {
+    val builder = new Http1Support(bufferSize, userAgent, es, osslContext, group)
+    builder.makeClient
+  }
+
+  private val Https: Scheme = "https".ci
+  private val Http: Scheme  = "http".ci
+}
+
 /** Provides basic HTTP1 pipeline building
-  *
-  * Also serves as a non-recycling [[ConnectionManager]] */
-final class Http1Support(bufferSize: Int,
-                            timeout: Duration,
+  */
+final private class Http1Support(bufferSize: Int,
                           userAgent: Option[`User-Agent`],
                                  es: ExecutorService,
                         osslContext: Option[SSLContext],
-                              group: Option[AsynchronousChannelGroup])
-  extends ConnectionBuilder with ConnectionManager
-{
+                              group: Option[AsynchronousChannelGroup]) {
   import Http1Support._
 
   private val ec = ExecutionContext.fromExecutorService(es)
-
   private val sslContext = osslContext.getOrElse(bits.sslContext)
   private val connectionManager = new ClientChannelFactory(bufferSize, group.orNull)
-
-  /** Get a connection to the provided address
-    * @param request [[Request]] to connect too
-    * @param fresh if the client should force a new connection
-    * @return a Future with the connected [[BlazeClientStage]] of a blaze pipeline
-    */
-  override def getClient(request: Request, fresh: Boolean): Task[BlazeClientStage] =
-    makeClient(request)
-
-  /** Free resources associated with this client factory */
-  override def shutdown(): Task[Unit] = Task.now(())
 
 ////////////////////////////////////////////////////
 
@@ -69,9 +76,18 @@ final class Http1Support(bufferSize: Int,
   }
 
   private def buildStages(uri: Uri): (LeafBuilder[ByteBuffer], BlazeClientStage) = {
-    val t = new Http1ClientStage(userAgent, timeout)(ec)
+    val t = new Http1ClientStage(userAgent, ec)
     val builder = LeafBuilder(t)
     uri match {
+      case Uri(Some(Https),Some(auth),_,_,_) =>
+        val eng = sslContext.createSSLEngine(auth.host.value, auth.port getOrElse 443)
+        eng.setUseClientMode(true)
+
+        val sslParams = eng.getSSLParameters
+        sslParams.setEndpointIdentificationAlgorithm("HTTPS")
+        eng.setSSLParameters(sslParams)
+
+        (builder.prepend(new SSLStage(eng)),t)
       case Uri(Some(Https),_,_,_,_) =>
         val eng = sslContext.createSSLEngine()
         eng.setUseClientMode(true)
@@ -92,7 +108,3 @@ final class Http1Support(bufferSize: Int,
   }
 }
 
-private object Http1Support {
-  private val Https: Scheme = "https".ci
-  private val Http: Scheme  = "http".ci
-}