+++ blaze-client/src/main/scala/org/http4s/client/blaze/PooledHttp1Client.scala
@@ -14,14 +14,15 @@ object PooledHttp1Client {
 
   /** Construct a new PooledHttp1Client */
   def apply(maxPooledConnections: Int = 10,
-                         timeout: Duration = DefaultTimeout,
-                       userAgent: Option[`User-Agent`] = DefaultUserAgent,
-                      bufferSize: Int = DefaultBufferSize,
-                        executor: ExecutorService = ClientDefaultEC,
+                     idleTimeout: Duration = bits.DefaultTimeout,
+                  requestTimeout: Duration = Duration.Inf,
+                       userAgent: Option[`User-Agent`] = bits.DefaultUserAgent,
+                      bufferSize: Int = bits.DefaultBufferSize,
+                        executor: ExecutorService = bits.ClientDefaultEC,
                       sslContext: Option[SSLContext] = None,
                            group: Option[AsynchronousChannelGroup] = None) = {
-    val http1 = new Http1Support(bufferSize, timeout, userAgent, executor, sslContext, group)
-    val pool = new PoolManager(maxPooledConnections, http1)
-    new BlazeClient(pool)
+    val http1 = Http1Support(bufferSize, userAgent, executor, sslContext, group)
+    val pool = ConnectionManager.pool(maxPooledConnections, http1)
+    new BlazeClient(pool, idleTimeout, requestTimeout)
   }
 }