+++ server/src/main/scala/org/http4s/server/Router.scala
@@ -1,24 +1,38 @@
-package org.http4s.server
-
-import org.http4s.server.middleware.URITranslation
-import org.http4s.{Status, Response, Request}
-
-import scalaz.concurrent.Task
+package org.http4s
+package server
 
 object Router {
-  // TODO A more efficient implementation does not require much imagination
-  def apply(mappings: (String, HttpService)*): HttpService = {
-    val table = mappings.sortBy(_._1.length).reverse.map { case (prefix, service) =>
-      val transformed =
-        if (prefix.isEmpty || prefix == "/") service
-        else URITranslation.translateRoot(prefix)(service)
-      prefix -> transformed
+
+  import Service.{withFallback => fallback}
+  import middleware.URITranslation.{translateRoot => translate}
+
+  /**
+    * Defines an HttpService based on list of mappings.
+    * @see define
+    */
+  def apply(mappings: (String, HttpService)*): HttpService = define(mappings:_*)()
+
+  /**
+    * Defines an HttpService based on list of mappings,
+    * a default Service to be used when none in the list match incomming requests,
+    * and an implicit Fallthrough which decides whether a request was matched.
+    *
+    * The mappings are processed in descending order (longest first) of prefix length.
+    */
+  def define(mappings: (String, HttpService)*)
+            (default: HttpService = HttpService.empty)
+            (implicit fallthrough: Fallthrough[Response]): HttpService =
+    mappings.sortBy(_._1.length).foldLeft(default) {
+      case (acc, (prefix, service)) =>
+        if (prefix.isEmpty || prefix == "/") fallback(acc)(service)(fallthrough)
+        else HttpService.lift {
+          req => (
+            if (req.pathInfo.startsWith(prefix))
+              fallback(acc)(translate(prefix)(service))(fallthrough)
+            else
+              acc
+          ) (req)
+        }
     }
 
-    Service.lift { req =>
-      table.find { case (prefix, _) =>
-        req.pathInfo.startsWith(prefix)
-      }.fold(HttpService.empty)(_._2)(req)
-    }
-  }
 }