+++ server/src/main/scala/org/http4s/server/syntax/Service.scala
@@ -1,0 +1,11 @@
+package org.http4s
+package server
+
+package object syntax {
+
+  final implicit class ServiceOps[A, B: Fallthrough](val service: Service[A, B]) {
+    def ||    (fallback: Service[A, B]) = orElse(fallback)
+    def orElse(fallback: Service[A, B]) = Service.withFallback(fallback)(service)
+  }
+
+}